Refactor worker into a sync task
This is needed for future features
